Make all the tools generic
